Fix: Extract logic into splitAmountAndFee util
